LayoutTests/media/video-error-abort-expected.txt

TEST(!video.error) OK
EVENT(loadstart)
EVENT(abort) TEST(video.error.code == MediaError.MEDIA_ERR_ABORTED) OK
EVENT(loadstart)
EVENT(load) TEST(!video.error) OK
END OF TEST
